Mauro: 4.3 Gestión de consejos territoriales. Falta eliminar seleccionados (pag57) [IT'S DONE]

+++ trunk/apps/frontend/modules/consejos_territoriales/actions/actions.class.php
@@ -16,7 +16,7 @@ class consejos_territorialesActions extends sfActions
 		if (is_numeric($this->paginaActual)) {
 			$this->getUser()->setAttribute($this->getModuleName().'_nowpage', $this->paginaActual);// recordar pagina actual
 		}
-  	    $this->pager = new sfDoctrinePager('ConsejoTerritorial', 10);
+  	    $this->pager = new sfDoctrinePager('ConsejoTerritorial', 4);
 		$this->pager->getQuery()
 		->from('ConsejoTerritorial ct')
 		->leftJoin('ct.UsuarioConsejoTerritorial uct')
@@ -68,18 +68,41 @@ class consejos_territorialesActions extends sfActions
 
   public function executeDelete(sfWebRequest $request)
   {
-    $request->checkCSRFProtection();
+  	
+  	$this->processSelectedRecords($request, 'baja');
 
-    $this->forward404Unless($consejos_territoriales = Doctrine::getTable('ConsejoTerritorial')->find($request->getParameter('id')), sprintf('Object consejos_territoriales does not exist (%s).', $request->getParameter('id')));
-    
-    sfLoader::loadHelpers('Security'); // para usar el helper
-	if (!validate_action('baja')) $this->redirect('seguridad/restringuido');
-    
-    $consejos_territoriales->delete();
-
-		$this->getUser()->setFlash('notice', "El registro ha sido eliminado del sistema");
-    $this->redirect('consejos_territoriales/index');
   }
+  protected function processSelectedRecords(sfWebRequest $request, $accion)
+  {
+  	$toProcess = $request->getParameter('id');
+  	
+  	if (!empty($toProcess)) {
+  		$request->checkCSRFProtection();
+  		
+  		$IDs = is_array($toProcess) ? $toProcess : array($toProcess);
+  		
+  		foreach ($IDs as $id) {
+  			    $this->forward404Unless($consejos_territoriales = Doctrine::getTable('ConsejoTerritorial')->find($id), sprintf('Object documentacion_grupo does not exist (%s).', $id));
+  			
+  			    sfLoader::loadHelpers('Security');
+				if (!validate_action($accion)) $this->redirect('seguridad/restringuido');
+
+				if ($accion == 'publicar') {
+//						$documentacion_grupo->setEstado('publicado');
+//						$documentacion_grupo->save();
+//			
+//						ServiceNotificacion::send('creacion', 'Grupo', $documentacion_grupo->getId(), $documentacion_grupo->getNombre(),'',$documentacion_grupo->getGrupoTrabajoId());	
+				} else {
+					$consejos_territoriales->delete();
+
+					$this->getUser()->setFlash('notice', "El registro ha sido eliminado del sistema");
+				}		
+  		}
+  	}
+  $this->redirect('consejos_territoriales/index');
+
+  }
+
 
   protected function processForm(sfWebRequest $request, sfForm $form, $accion='')
   {
